refactor(desktop-tauri): split workflow rules and refine runtime command flow
