--- apps/desktop/src-tauri/crates/host-application/src/app_service/workflow_rules.rs
@@ -1,695 +1,0 @@
-use anyhow::{anyhow, Result};
-use host_domain::{
-    AgentWorkflowState, AgentWorkflows, CreateTaskInput, IssueType, PlanSubtaskInput,
-    QaWorkflowVerdict, TaskAction, TaskCard, TaskStatus, UpdateTaskPatch,
-};
-
-pub(crate) fn default_qa_required_for_issue_type(issue_type: &IssueType) -> bool {
-    matches!(
-        issue_type,
-        IssueType::Epic | IssueType::Feature | IssueType::Task | IssueType::Bug
-    )
-}
-
-pub(crate) fn is_open_state(status: &TaskStatus) -> bool {
-    !matches!(status, TaskStatus::Closed | TaskStatus::Deferred)
-}
-
-fn can_skip_spec_and_planning(task: &TaskCard) -> bool {
-    matches!(task.issue_type, IssueType::Task | IssueType::Bug)
-}
-
-pub(crate) fn derive_agent_workflows(task: &TaskCard) -> AgentWorkflows {
-    let is_feature_epic = matches!(task.issue_type, IssueType::Feature | IssueType::Epic);
-    let is_task_bug = matches!(task.issue_type, IssueType::Task | IssueType::Bug);
-    let qa_required = task.ai_review_enabled;
-    let is_closed = task.status == TaskStatus::Closed;
-    let is_ready_for_dev_or_later = matches!(
-        task.status,
-        TaskStatus::ReadyForDev
-            | TaskStatus::InProgress
-            | TaskStatus::Blocked
-            | TaskStatus::AiReview
-            | TaskStatus::HumanReview
-    );
-    let is_planner_feature_epic_status =
-        task.status == TaskStatus::SpecReady || is_ready_for_dev_or_later;
-
-    let spec_required = is_feature_epic;
-    let spec_can_skip = !spec_required;
-    let spec_available = !is_closed;
-    let spec_completed = task.document_summary.spec.has;
-
-    let planner_required = is_feature_epic;
-    let planner_can_skip = !planner_required;
-    let planner_available = if is_closed {
-        false
-    } else if is_task_bug {
-        true
-    } else if is_feature_epic {
-        is_planner_feature_epic_status
-    } else {
-        false
-    };
-    let planner_completed = task.document_summary.plan.has;
-
-    let builder_available = if is_closed {
-        false
-    } else if is_task_bug {
-        true
-    } else if is_feature_epic {
-        is_ready_for_dev_or_later
-    } else {
-        false
-    };
-    let builder_completed = matches!(
-        task.status,
-        TaskStatus::AiReview | TaskStatus::HumanReview | TaskStatus::Closed
-    );
-
-    let qa_available = if is_closed {
-        false
-    } else {
-        task.status == TaskStatus::AiReview
-    };
-    let qa_completed = task.document_summary.qa_report.verdict == QaWorkflowVerdict::Approved;
-
-    AgentWorkflows {
-        spec: AgentWorkflowState {
-            required: spec_required,
-            can_skip: spec_can_skip,
-            available: spec_available,
-            completed: spec_completed,
-        },
-        planner: AgentWorkflowState {
-            required: planner_required,
-            can_skip: planner_can_skip,
-            available: planner_available,
-            completed: planner_completed,
-        },
-        builder: AgentWorkflowState {
-            required: true,
-            can_skip: false,
-            available: builder_available,
-            completed: builder_completed,
-        },
-        qa: AgentWorkflowState {
-            required: qa_required,
-            can_skip: !qa_required,
-            available: qa_available,
-            completed: qa_completed,
-        },
-    }
-}
-
-pub(crate) fn allows_transition(task: &TaskCard, from: &TaskStatus, to: &TaskStatus) -> bool {
-    if from == to {
-        return true;
-    }
-
-    match from {
-        TaskStatus::Open => {
-            if can_skip_spec_and_planning(task) {
-                matches!(
-                    to,
-                    TaskStatus::SpecReady
-                        | TaskStatus::ReadyForDev
-                        | TaskStatus::InProgress
-                        | TaskStatus::Deferred
-                )
-            } else {
-                matches!(to, TaskStatus::SpecReady | TaskStatus::Deferred)
-            }
-        }
-        TaskStatus::SpecReady => {
-            if can_skip_spec_and_planning(task) {
-                matches!(
-                    to,
-                    TaskStatus::ReadyForDev | TaskStatus::InProgress | TaskStatus::Deferred
-                )
-            } else {
-                matches!(to, TaskStatus::ReadyForDev | TaskStatus::Deferred)
-            }
-        }
-        TaskStatus::ReadyForDev => matches!(to, TaskStatus::InProgress | TaskStatus::Deferred),
-        TaskStatus::InProgress => {
-            matches!(
-                to,
-                TaskStatus::Blocked
-                    | TaskStatus::AiReview
-                    | TaskStatus::HumanReview
-                    | TaskStatus::Deferred
-            )
-        }
-        TaskStatus::Blocked => matches!(to, TaskStatus::InProgress | TaskStatus::Deferred),
-        TaskStatus::AiReview => matches!(
-            to,
-            TaskStatus::InProgress | TaskStatus::HumanReview | TaskStatus::Deferred
-        ),
-        TaskStatus::HumanReview => matches!(
-            to,
-            TaskStatus::InProgress | TaskStatus::Closed | TaskStatus::Deferred
-        ),
-        TaskStatus::Deferred => matches!(to, TaskStatus::Open),
-        TaskStatus::Closed => false,
-    }
-}
-
-pub(crate) fn validate_transition(
-    task: &TaskCard,
-    all_tasks: &[TaskCard],
-    from: &TaskStatus,
-    to: &TaskStatus,
-) -> Result<()> {
-    if !allows_transition(task, from, to) {
-        return Err(anyhow!(
-            "Transition not allowed for {} ({}): {} -> {}",
-            task.id,
-            task.issue_type.as_cli_value(),
-            from.as_cli_value(),
-            to.as_cli_value()
-        ));
-    }
-
-    if *to == TaskStatus::Closed && task.issue_type == IssueType::Epic {
-        let blocking_subtasks = all_tasks.iter().filter(|candidate| {
-            candidate.parent_id.as_deref() == Some(task.id.as_str())
-                && !matches!(candidate.status, TaskStatus::Closed | TaskStatus::Deferred)
-        });
-
-        if let Some(first_blocking) = blocking_subtasks.take(1).next() {
-            return Err(anyhow!(
-                "Epic cannot be completed while direct subtask {} is still active.",
-                first_blocking.id
-            ));
-        }
-    }
-
-    Ok(())
-}
-
-fn find_task<'a>(tasks: &'a [TaskCard], task_id: &str) -> Result<&'a TaskCard> {
-    tasks
-        .iter()
-        .find(|task| task.id == task_id)
-        .ok_or_else(|| anyhow!("Task not found: {task_id}"))
-}
-
-pub(crate) fn validate_parent_relationships_for_create(
-    tasks: &[TaskCard],
-    input: &CreateTaskInput,
-) -> Result<()> {
-    let parent_id = input
-        .parent_id
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty());
-
-    if input.issue_type == IssueType::Epic && parent_id.is_some() {
-        return Err(anyhow!("Epics cannot be created as subtasks."));
-    }
-
-    if let Some(parent_id) = parent_id {
-        let parent = find_task(tasks, parent_id)?;
-        if parent.issue_type != IssueType::Epic {
-            return Err(anyhow!("Only epics can have subtasks."));
-        }
-        if parent.parent_id.is_some() {
-            return Err(anyhow!("Subtask depth is limited to one level."));
-        }
-    }
-
-    Ok(())
-}
-
-pub(crate) fn validate_parent_relationships_for_update(
-    tasks: &[TaskCard],
-    current: &TaskCard,
-    patch: &UpdateTaskPatch,
-) -> Result<()> {
-    let next_issue_type = patch
-        .issue_type
-        .clone()
-        .unwrap_or_else(|| current.issue_type.clone());
-
-    let next_parent_id = match patch.parent_id.as_deref() {
-        Some(value) => {
-            let trimmed = value.trim();
-            if trimmed.is_empty() {
-                None
-            } else {
-                Some(trimmed)
-            }
-        }
-        None => current.parent_id.as_deref(),
-    };
-
-    if next_issue_type == IssueType::Epic && next_parent_id.is_some() {
-        return Err(anyhow!("Epics cannot be converted to subtasks."));
-    }
-
-    let has_direct_subtasks = tasks
-        .iter()
-        .any(|task| task.parent_id.as_deref() == Some(current.id.as_str()));
-    if has_direct_subtasks && next_parent_id.is_some() {
-        return Err(anyhow!("Tasks with subtasks cannot become subtasks."));
-    }
-
-    if has_direct_subtasks && next_issue_type != IssueType::Epic {
-        return Err(anyhow!("Only epics can have subtasks."));
-    }
-
-    if let Some(parent_id) = next_parent_id {
-        let parent = find_task(tasks, parent_id)?;
-        if parent.issue_type != IssueType::Epic {
-            return Err(anyhow!("Only epics can be selected as parents."));
-        }
-        if parent.parent_id.is_some() {
-            return Err(anyhow!("Subtask depth is limited to one level."));
-        }
-    }
-
-    Ok(())
-}
-
-pub(crate) fn normalize_required_markdown(markdown: &str, document_label: &str) -> Result<String> {
-    let trimmed = markdown.trim();
-    if trimmed.is_empty() {
-        return Err(anyhow!("{document_label} markdown cannot be empty."));
-    }
-    Ok(trimmed.to_string())
-}
-
-pub(crate) fn can_set_spec_from_status(status: &TaskStatus) -> bool {
-    matches!(status, TaskStatus::Open | TaskStatus::SpecReady)
-}
-
-pub(crate) fn can_set_plan(task: &TaskCard) -> bool {
-    match task.issue_type {
-        IssueType::Epic | IssueType::Feature => {
-            matches!(task.status, TaskStatus::SpecReady | TaskStatus::ReadyForDev)
-        }
-        IssueType::Task | IssueType::Bug => matches!(
-            task.status,
-            TaskStatus::Open | TaskStatus::SpecReady | TaskStatus::ReadyForDev
-        ),
-    }
-}
-
-pub(crate) fn can_replace_epic_subtask_status(status: &TaskStatus) -> bool {
-    matches!(
-        status,
-        TaskStatus::Open | TaskStatus::SpecReady | TaskStatus::ReadyForDev
-    )
-}
-
-pub(crate) fn derive_available_actions(task: &TaskCard, all_tasks: &[TaskCard]) -> Vec<TaskAction> {
-    let mut actions = vec![TaskAction::ViewDetails];
-
-    if can_set_spec_from_status(&task.status) {
-        actions.push(TaskAction::SetSpec);
-    }
-
-    if can_set_plan(task) {
-        actions.push(TaskAction::SetPlan);
-    }
-
-    if allows_transition(task, &task.status, &TaskStatus::InProgress) {
-        actions.push(TaskAction::BuildStart);
-    }
-
-    if matches!(
-        task.status,
-        TaskStatus::InProgress | TaskStatus::Blocked | TaskStatus::HumanReview
-    ) {
-        actions.push(TaskAction::OpenBuilder);
-    }
-
-    if task.parent_id.is_none() {
-        if task.status == TaskStatus::Deferred {
-            actions.push(TaskAction::ResumeDeferred);
-        } else if is_open_state(&task.status) {
-            actions.push(TaskAction::DeferIssue);
-        }
-    }
-
-    if task.status == TaskStatus::HumanReview {
-        actions.push(TaskAction::HumanRequestChanges);
-    }
-
-    if validate_transition(task, all_tasks, &task.status, &TaskStatus::Closed).is_ok() {
-        actions.push(TaskAction::HumanApprove);
-    }
-
-    actions
-}
-
-pub(crate) fn validate_plan_subtask_rules(
-    task: &TaskCard,
-    all_tasks: &[TaskCard],
-    plan_subtasks: &[CreateTaskInput],
-) -> Result<()> {
-    if task.issue_type != IssueType::Epic {
-        if !plan_subtasks.is_empty() {
-            return Err(anyhow!(
-                "Only epics can receive subtask proposals during planning."
-            ));
-        }
-        return Ok(());
-    }
-
-    let has_direct_subtasks = all_tasks
-        .iter()
-        .any(|entry| entry.parent_id.as_deref() == Some(task.id.as_str()));
-    if !has_direct_subtasks && plan_subtasks.is_empty() {
-        return Err(anyhow!(
-            "Epic plans must provide at least one direct subtask proposal."
-        ));
-    }
-
-    Ok(())
-}
-
-pub(crate) fn normalize_title_key(title: &str) -> String {
-    title.trim().to_ascii_lowercase()
-}
-
-pub(crate) fn normalize_subtask_plan_inputs(
-    inputs: Vec<PlanSubtaskInput>,
-) -> Result<Vec<CreateTaskInput>> {
-    let mut normalized = Vec::with_capacity(inputs.len());
-    for entry in inputs {
-        let title = entry.title.trim().to_string();
-        if title.is_empty() {
-            return Err(anyhow!("Subtask proposals require a non-empty title."));
-        }
-
-        let issue_type = entry.issue_type.unwrap_or(IssueType::Task);
-        if issue_type == IssueType::Epic {
-            return Err(anyhow!(
-                "Epic subtasks are not allowed. Subtask hierarchy depth is limited to one level."
-            ));
-        }
-
-        let priority = entry.priority.unwrap_or(2).clamp(0, 4);
-        let description = entry.description.and_then(|value| {
-            let trimmed = value.trim();
-            if trimmed.is_empty() {
-                None
-            } else {
-                Some(trimmed.to_string())
-            }
-        });
-
-        normalized.push(CreateTaskInput {
-            title,
-            issue_type: issue_type.clone(),
-            priority,
-            description,
-            acceptance_criteria: None,
-            labels: None,
-            ai_review_enabled: Some(default_qa_required_for_issue_type(&issue_type)),
-            parent_id: None,
-        });
-    }
-    Ok(normalized)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        allows_transition, can_replace_epic_subtask_status, can_set_plan, can_set_spec_from_status,
-        derive_agent_workflows, derive_available_actions, normalize_subtask_plan_inputs,
-        normalize_title_key,
-    };
-    use crate::app_service::test_support::make_task;
-    use host_domain::{PlanSubtaskInput, QaWorkflowVerdict, TaskAction, TaskStatus};
-    use serde::Deserialize;
-    use std::collections::BTreeMap;
-
-    #[derive(Debug, Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    struct WorkflowContractFixture {
-        statuses: Vec<String>,
-        transitions: BTreeMap<String, BTreeMap<String, Vec<String>>>,
-        set_spec_allowed_statuses: Vec<String>,
-        set_plan_allowed_statuses: BTreeMap<String, Vec<String>>,
-        epic_subtask_replacement_allowed_statuses: Vec<String>,
-    }
-
-    fn load_workflow_contract_fixture() -> WorkflowContractFixture {
-        let raw =
-            include_str!("../../../../../../../docs/contracts/workflow-contract-fixture.json");
-        serde_json::from_str(raw).expect("workflow contract fixture must parse")
-    }
-
-    fn parse_status(value: &str) -> TaskStatus {
-        TaskStatus::from_cli_value(value)
-            .unwrap_or_else(|| panic!("unknown fixture status: {value}"))
-    }
-
-    #[test]
-    fn module_normalize_title_key_is_case_insensitive_and_trimmed() {
-        assert_eq!(normalize_title_key("  Build Runtime  "), "build runtime");
-        assert_eq!(normalize_title_key("BUILD runtime"), "build runtime");
-    }
-
-    #[test]
-    fn module_derive_available_actions_exposes_resume_for_deferred_task() {
-        let deferred = make_task("task-1", "task", TaskStatus::Deferred);
-
-        let actions = derive_available_actions(&deferred, &[deferred.clone()]);
-
-        assert!(actions.contains(&TaskAction::ResumeDeferred));
-    }
-
-    #[test]
-    fn module_normalize_subtask_plan_inputs_rejects_empty_title() {
-        let inputs = vec![PlanSubtaskInput {
-            title: "   ".to_string(),
-            issue_type: None,
-            priority: None,
-            description: None,
-        }];
-
-        let error =
-            normalize_subtask_plan_inputs(inputs).expect_err("empty title should be rejected");
-
-        assert!(error.to_string().contains("title"));
-    }
-
-    #[test]
-    fn module_derive_agent_workflows_spec_availability_is_false_only_when_closed() {
-        let mut task = make_task("task-1", "feature", TaskStatus::Open);
-        for status in [
-            TaskStatus::Open,
-            TaskStatus::SpecReady,
-            TaskStatus::ReadyForDev,
-            TaskStatus::InProgress,
-            TaskStatus::Blocked,
-            TaskStatus::AiReview,
-            TaskStatus::HumanReview,
-            TaskStatus::Deferred,
-        ] {
-            task.status = status;
-            let workflows = derive_agent_workflows(&task);
-            assert!(workflows.spec.available);
-        }
-
-        task.status = TaskStatus::Closed;
-        let workflows = derive_agent_workflows(&task);
-        assert!(!workflows.spec.available);
-    }
-
-    #[test]
-    fn module_derive_agent_workflows_planner_and_builder_availability_matrix() {
-        let mut task = make_task("task-1", "task", TaskStatus::Open);
-        for status in [
-            TaskStatus::Open,
-            TaskStatus::SpecReady,
-            TaskStatus::ReadyForDev,
-            TaskStatus::InProgress,
-            TaskStatus::Blocked,
-            TaskStatus::AiReview,
-            TaskStatus::HumanReview,
-            TaskStatus::Deferred,
-        ] {
-            task.status = status;
-            let workflows = derive_agent_workflows(&task);
-            assert!(workflows.planner.available);
-            assert!(workflows.builder.available);
-        }
-        task.status = TaskStatus::Closed;
-        let workflows = derive_agent_workflows(&task);
-        assert!(!workflows.planner.available);
-        assert!(!workflows.builder.available);
-
-        let mut feature = make_task("task-2", "feature", TaskStatus::Open);
-        let feature_open = derive_agent_workflows(&feature);
-        assert!(!feature_open.planner.available);
-        assert!(!feature_open.builder.available);
-
-        feature.status = TaskStatus::SpecReady;
-        let feature_spec_ready = derive_agent_workflows(&feature);
-        assert!(feature_spec_ready.planner.available);
-        assert!(!feature_spec_ready.builder.available);
-
-        feature.status = TaskStatus::ReadyForDev;
-        let feature_ready_for_dev = derive_agent_workflows(&feature);
-        assert!(feature_ready_for_dev.planner.available);
-        assert!(feature_ready_for_dev.builder.available);
-
-        feature.status = TaskStatus::HumanReview;
-        let feature_human_review = derive_agent_workflows(&feature);
-        assert!(feature_human_review.planner.available);
-        assert!(feature_human_review.builder.available);
-
-        feature.status = TaskStatus::Closed;
-        let feature_closed = derive_agent_workflows(&feature);
-        assert!(!feature_closed.planner.available);
-        assert!(!feature_closed.builder.available);
-    }
-
-    #[test]
-    fn module_derive_agent_workflows_qa_flags_and_completion_follow_payload() {
-        let mut task = make_task("task-1", "task", TaskStatus::AiReview);
-        task.ai_review_enabled = true;
-        let required = derive_agent_workflows(&task);
-        assert!(required.qa.required);
-        assert!(!required.qa.can_skip);
-        assert!(required.qa.available);
-        assert!(!required.qa.completed);
-
-        task.ai_review_enabled = false;
-        let optional = derive_agent_workflows(&task);
-        assert!(!optional.qa.required);
-        assert!(optional.qa.can_skip);
-        assert!(optional.qa.available);
-
-        task.document_summary.qa_report.verdict = QaWorkflowVerdict::Rejected;
-        let rejected = derive_agent_workflows(&task);
-        assert!(!rejected.qa.completed);
-
-        task.document_summary.qa_report.verdict = QaWorkflowVerdict::NotReviewed;
-        let not_reviewed = derive_agent_workflows(&task);
-        assert!(!not_reviewed.qa.completed);
-
-        task.document_summary.qa_report.verdict = QaWorkflowVerdict::Approved;
-        let approved = derive_agent_workflows(&task);
-        assert!(approved.qa.completed);
-    }
-
-    #[test]
-    fn module_derive_agent_workflows_closed_precedence_and_reopen_recompute() {
-        let mut task = make_task("task-1", "feature", TaskStatus::InProgress);
-        let open_state = derive_agent_workflows(&task);
-        assert!(open_state.spec.available);
-        assert!(open_state.planner.available);
-        assert!(open_state.builder.available);
-        assert!(!open_state.qa.available);
-
-        task.status = TaskStatus::Closed;
-        let closed_state = derive_agent_workflows(&task);
-        assert!(!closed_state.spec.available);
-        assert!(!closed_state.planner.available);
-        assert!(!closed_state.builder.available);
-        assert!(!closed_state.qa.available);
-
-        task.status = TaskStatus::AiReview;
-        let reopened = derive_agent_workflows(&task);
-        assert!(reopened.spec.available);
-        assert!(reopened.planner.available);
-        assert!(reopened.builder.available);
-        assert!(reopened.qa.available);
-    }
-
-    #[test]
-    fn workflow_contract_transitions_match_fixture() {
-        let fixture = load_workflow_contract_fixture();
-        for issue_type in ["epic", "feature", "task", "bug"] {
-            let issue_transitions = fixture
-                .transitions
-                .get(issue_type)
-                .unwrap_or_else(|| panic!("missing transitions for issue type: {issue_type}"));
-            for from in &fixture.statuses {
-                let from_status = parse_status(from);
-                let task = make_task("fixture-task", issue_type, from_status.clone());
-                let expected_targets = issue_transitions
-                    .get(from)
-                    .unwrap_or_else(|| panic!("missing transitions for status: {from}"));
-
-                for to in &fixture.statuses {
-                    let to_status = parse_status(to);
-                    let expected_allowed = from == to || expected_targets.contains(to);
-                    let actual_allowed = allows_transition(&task, &from_status, &to_status);
-                    assert_eq!(
-                        actual_allowed, expected_allowed,
-                        "transition mismatch for issue_type={issue_type} {from}->{to}"
-                    );
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn workflow_contract_set_spec_statuses_match_fixture() {
-        let fixture = load_workflow_contract_fixture();
-        let expected = fixture
-            .set_spec_allowed_statuses
-            .iter()
-            .map(String::as_str)
-            .collect::<Vec<_>>();
-
-        for status in &fixture.statuses {
-            let parsed = parse_status(status);
-            let actual_allowed = can_set_spec_from_status(&parsed);
-            let expected_allowed = expected.contains(&status.as_str());
-            assert_eq!(
-                actual_allowed, expected_allowed,
-                "set_spec mismatch for status={status}"
-            );
-        }
-    }
-
-    #[test]
-    fn workflow_contract_set_plan_statuses_match_fixture() {
-        let fixture = load_workflow_contract_fixture();
-        for issue_type in ["epic", "feature", "task", "bug"] {
-            let expected_statuses = fixture
-                .set_plan_allowed_statuses
-                .get(issue_type)
-                .unwrap_or_else(|| panic!("missing set_plan statuses for issue_type={issue_type}"));
-
-            for status in &fixture.statuses {
-                let parsed_status = parse_status(status);
-                let task = make_task("fixture-task", issue_type, parsed_status);
-                let actual_allowed = can_set_plan(&task);
-                let expected_allowed = expected_statuses.contains(status);
-                assert_eq!(
-                    actual_allowed, expected_allowed,
-                    "set_plan mismatch for issue_type={issue_type} status={status}"
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn workflow_contract_epic_subtask_replacement_statuses_match_fixture() {
-        let fixture = load_workflow_contract_fixture();
-        let expected = fixture
-            .epic_subtask_replacement_allowed_statuses
-            .iter()
-            .map(String::as_str)
-            .collect::<Vec<_>>();
-
-        for status in &fixture.statuses {
-            let parsed = parse_status(status);
-            let actual_allowed = can_replace_epic_subtask_status(&parsed);
-            let expected_allowed = expected.contains(&status.as_str());
-            assert_eq!(
-                actual_allowed, expected_allowed,
-                "epic subtask replacement mismatch for status={status}"
-            );
-        }
-    }
-}
