+++ apps/desktop/src-tauri/crates/host-application/src/app_service/workflow_rules/mod.rs
@@ -1,0 +1,15 @@
+mod transitions;
+mod validators;
+
+pub(crate) use transitions::{
+    allows_transition, can_replace_epic_subtask_status, can_set_plan, can_set_spec_from_status,
+    default_qa_required_for_issue_type, derive_agent_workflows, is_open_state,
+};
+pub(crate) use validators::{
+    derive_available_actions, normalize_required_markdown, normalize_subtask_plan_inputs,
+    normalize_title_key, validate_parent_relationships_for_create,
+    validate_parent_relationships_for_update, validate_plan_subtask_rules, validate_transition,
+};
+
+#[cfg(test)]
+mod tests;
